+++ src/main/java/SouffleLint.java
@@ -1,0 +1,8 @@
+import java.util.List;
+
+public class SouffleLint {
+    List<SouffleLintContext> fragments;
+    SouffleLintRule rule;
+    String source;
+    String source_file;
+}