+++ src/main/java/SouffleLintContext.java
@@ -1,0 +1,7 @@
+public class SouffleLintContext {
+    String node_text;
+    SouffleLintContext context;
+    SouffleLintPoint start;
+    SouffleLintPoint end;
+
+}