+++ src/main/java/SouffleLintExample.java
@@ -1,0 +1,4 @@
+public class SouffleLintExample {
+    String before;
+    String after;
+}