+++ src/main/java/SouffleLintPoint.java
@@ -1,0 +1,4 @@
+public class SouffleLintPoint {
+    int row;
+    int column;
+}